Force SSL in production
* Vehicle seat_info Example in README
* Omniauth callback

+++ app/models/vehicle.rb
@@ -1,6 +1,6 @@
 class Vehicle < ActiveRecord::Base
   after_update :check_seats!
-  after_create :check_seats
+  after_create :check_seats!
 
   has_many :schedules
   has_many :orders
@@ -42,7 +42,10 @@ class Vehicle < ActiveRecord::Base
   #   ]
   # }
   private
-    def check_seat
+    def check_seats!
+      return unless self.seat_info_changed?
+      return unless self.seat_info[:seating] && self.seat_info[:seating].is_a?(Array)
+
       Seat.destroy_all(vehicle: self)
       # may need to handle orders
 
@@ -55,12 +58,5 @@ class Vehicle < ActiveRecord::Base
           i += 1
         end
       end
-    end
-
-    def check_seats!
-      return unless self.seat_info_changed?
-      return unless self.seat_info[:seating] && self.seat_info[:seating].is_a?(Array)
-
-      check_seat
     end
 end
